requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-datatables==0.7.2
odata-query==0.10.0
python-dateutil==2.9.0.post0
six==1.17.0
sly==0.4
sqlparse==0.5.3
tzdata==2025.2
utils==1.0.2
